Remove completely references of payment, legacy (wiki) solutions, printonly exams
Oral exam deposits will not be used at Edinburgh, as there are not enough
oral exams to justify its maintenance. References to VSETH Legacy
solutions and Exams/Solutions that can only be printed are also very
specific to the usecase of ETH Zurich.

They were removed to slim down backend and ease maintenance at Edinburgh.

This commit includes manual changes in the database migration for the
scoreboard module (as it contained manually written references to the
legacy answer system). The database must be rebuilt from scratch.