Use full width in login form card, use inline button
The login form now uses the full width of the <Paper>, because otherwise
the side margins were way too large.

As a consequence, the login button looked very plain, so it is now an
inline button within the text field (achieved with <ActionIcon>). It
still has type="submit" so that pressing enter in the text field works.