Add raw.githubusercontent to CSP (only under CompSoc org), use nofill signet logo
The signet specified in config.js was the the filled version, which when
white-filled (as it is on the mobile header) looks bad. This commit
changes to use the no-fill version, which unfortunately is not available
on the comp-soc.com domain, so we add raw.githubusercontent.com to the
CSP list (condition on the repo owner being compsoc-edinburgh), and use
the file from there.