Decode the JWT token on client-side to check auth status
Maintaining another cookie for token_expiry was not very clean. Instead,
the client will now decode the JWT auth token to check if it is expired
when determining if the user is logged in. Functionally this is the same
as previous but it has a cleaner interface with the backend.