Make admin generate JWT keypair manually, not in cinit
This is because if it is in cinit, each instance of the built image
will generate a new keypair, so if users go between instances (e.g.
using Kubernetes load-balancing), they will lose authentication.

Plus making admins generate keypairs manually is beneficial in that
the frequent container restarts while development doesn't lose auth.