Fix null constraint database error when logging in twice
There was a bug in the way the verification code was inserted into the
database. The .save() method which was used for automaticaly creating or
updating a row in the verification code table passed a NULL value for
'created_at' if the row was already present, because a value for it was
not supplied in code.

This could be caused when a user generates a verification code, does not
use it (because using it deletes the row), and generates another code
(after the 1 hour expiry).

We want to essentially only update the dirty fields (generated code)
when updating, but we also want to set the 'created_at' field when
creating a new row. The .save() method isn't smart enough to distinguish
these cases, so we change the code to use .update_or_create() instead
(which has support for this behaviour from Django 4.2).

Because the 'auto_now_add' option isn't supported when using
.update_or_create (I think), and because the automagic behaviour of it
adds an unnecessary level of indirection, the commit also replaces this
field option with a 'default' option with equivalent values. A migration
is included to update the database.