Fix backend default logger handler not receiving django logs
It appears that Django logs don't propagate further than the logger
with name 'django' (even the docs refer to the 'django' logger as
the root logger). This meant that the previously configured log
handler at the real root wasn't receiving any of the django events,
making it hard to debug when DEBUG was False.

This commit fixes it -- all django logs above WARNING level are
output to stdout in production too.