Simplify category card (hide answered exam count)
The "number of exams with at least one answer" wasn't very useful
to know nor was it clear UI wise, so it is removed from the
category card in the frontend.

The backend has some comments added to clarify to future developers
what the fields in the category meta view mean.