Vite Migration: Allow data: as font-src
One of the KaTeX fonts is embedded as a base64-encoded string in the CSS file when using Vite.
Our current font-src CSP does not allow this.
Theoretically this does reduce security a little bit, but since we already allow "unsafe-inline" for style-src and "unsafe-eval" for script-src, it doesn't matter at all.