Reduce Mantine LoadingOverlay default opacity, increase transition
With the move to Mantine's LoadingOverlay, the 1s animation on opacity
was lost (and the color changed from gray to white). This meant that
loading cached/quick pages showed a white flash which was quite
distracting. This commit creates a wrapper component and sets the
default opacity to 0.1, color to gray, and transition to 0.5s to mimic
the old behavior.