Allow non-admins to upload new exams
This encourages users to upload exams, and although the modqueue step
is still necessary (to cut up the PDF and make it public), it makes it
easier for users to contribute and grow the database. The status quo in
comparison requires users to contact admins by email/Discord to request
that they upload an exam.