Remove IP from email as the production env doens't have access to client IP
The production environment runs as a Kubernetes service behind Traefik
(configured with Ingress). Traefik is configured to strip the client IP
from the request headers, so the IP address is not available to the
application. This means there's no point in sending the email with the
IP address, as it will always be the local pod IP.