Fix tight margins in answer list when viewing a user profile
After b050388, <AnswerComponent> no longer has a top and bottom margin.
The margin was removed to create cohesion between an answer and its
associated PDF question, but this change also affected the user profile
page where answers are listed.

This commit modifies the map() function in <UserAnswers> to add a bottom
margin to each answer component.