Fix login for ed.ac.uk accounts broken after eeb2bc4
Since the auth backend checked for the existence of a claimed user
using both the username and email, logins failed for after the email
for logins was changed to sms.ed.ac.uk while the database still retained
the old ed.ac.uk emails.

Since the username should be enough to uniquely identify a user, the
email check was removed.