Lengthen auth token validity from 4 to 12 weeks
There's no particular reason we should have a 4
week limit, and this is also quite short given the
length of a term and the student lifecycle.

It's also very unlikely for a user to become
ineligible for access in 3 months if they could
access the system before, so it was just additional
hassle for users to have to re-authenticate.