Fix scroll to answer
Currently opening an answer via a direct link (i.e. with the hash in the URL) results in strange behavior where the element is being repeatedly scrolled into view. I believe this was caused by an unnecessary rerender of the HashLocationHandler component. Also, if the component gets rerendered, cleanup isn't being done properly. Previously the exam page would also jump to the location itself, now that's only the responsibility of HashLocationHandler.

Signed-off-by: Thomas Gassmann <tgassmann@student.ethz.ch>