Fix BottomHeader (the only header in our deployment) having malformed color toggle
Because of the placement of the ColorSchemeToggle component, the alignment in the
header was broken. This is fixed.

Further, in upstream the toggle is located in the TopHeader which has a constant
dark background regardless of the effective color scheme. In our deployment, we
put this toggle in BottomHeader (since we don't have a TopHeader to simplify nav),
which changes background color (also included in this commit) -- so the toggle icon
also needed some changes to properly change color depending on the scheme.