Fix #5: padlock overlay on CategoryCard having too high zIndex
LoadingOverlay by default has a zIndex of 400. The navigation bar
for example has --mantine-z-index-app which is 100. (Which makes
sense since if we're doing a full-page LoadingOverlay we want it
to be shown over the navigation bar I think... although unsure now).

Anyway, for now this commit reduces the CategoryCard overlays to
have a zIndex of just 2, so it draws over the card itself but under
the navigation bar and other elements.