Reduce layout shift when navigating to category page with cached exams
When navigating to a category page with the exams cached, the page
exam section would shift downwards for a few seconds because of a
Loader component. This was jarring and also prevented you from clicking
on an exam quickly, despite the data already being there.

This commit moves the loader to an absolute position just to the top
right of the exam list, akin to how it is done in the home page.
(Although, unlike the homepage, the exam data is only cached for
SPA navigations and does not persist across page reloads -- maybe
it's worth investigating using localstorage cache for exam data too,
but at that point I feel like we're abusing browser local storage when
there are other solutions out there)