Fix Deploy stage not running if either of Typecheck/UTs were skipped
This happens due to https://github.com/actions/runner/issues/2205,
where if there is a chain like:
A -> B -> C
where if you want the chain to continue after A is skipped (not failed!)
you need to insert a `!failure() && !canceled()` as a condition for every
subsequent job (i.e. both for B and C).

This seems like a bug because intuitively you'd only set it for B,
but the issue seems open with no actions taken, so we'll just follow the
workaround of also setting the conditional for C too.