Fix `mc config` being deprecated and unavailable in CI
In CI we use the latest minio images and client (unlike our
development environment docker compose file, which is pinned
to a version from 2023). Since a few days ago, the minio
client has removed the `mc config` command and repalced it
with `mc alias`. This commit fixes the CI to use the new
command.

As a side note, the reason we rely on the latest minio image
instead of pinning it in the CI too, is that we use a third-
party minio image (maragudk/minio-ci) to work around some
limitations in the official image. This third-pary image
doesn't have the old release from 2023.