Fix category page BI course data disappearing during dev-mode HMR
This occured because the hot module reload causes the network request to
fire with an existing data -- which leads to loading=true and data!=undefined.

For some reason, loading never resets to false after HMR. To work
around this, we don't use the loading state for the course data, just
whether data != undefined or not.