Remove unnecessary re-render on CommentSectionComponent when URL has comment param
`useEffect` runs after render. Thus, calling setState inside of it (when there is a
`comment` URL param) causes a re-render of the component.
This commit fixes that by statelessly calculating `expanded` as the logical OR of
the URL's desired expansion state and the user-toggled expansion state. The
implementation is then pure and does not need to be enclosed in `useEffect` anymore.