Use children instead of deprecated component prop for frontend route content
Using direct children (or via children prop) is the recommended way forward
to allow migration to React Router v6/7.

This change required amendment to the props in LoginPage (to accurately
reflect the type of isHome) and to HashLocationHandler (to use React hooks
for getting the hash instead of via props).