Fix navbar and body shifting when page is long enough to be scrollable
This fixes the navbar and the body shifting to the left on the home page
when the category cards load. The issue was also seen when navigating from
a short page like /search to a long page like the home page. One could
argue that it's a browser feature that users expect, but I argue that
this is a case where overriding the browser's behavior is worth it for
a smoother feel.