Move 'Load more comments' button in CommentSectionComponent to a comment-like box
This is more natural UI to convey that comments are collapsed.
It also helps when adding a new comment on a collapsed thread,
because previously you had no indicator that there were unseen
comments that were being replied to. Now, it is visible that
there are collapsed comments above your new comment.