Fix regression in adding cuts for already finished exams (request from LUK)
In !520 we assumed that admins probably will not want to accidentally
edit or add cuts to exams that are already marked as "Finished Cuts".

This was a wrong assumption to make, as LUK frequently fixes bad cuts
etc through clicking this button even for exams already marked as finished.

Thus we revert some of the functionality to how it was.

Additionally, !520 introduced a visual regression where the 'snap'
checkbox would remain visible if you released a claim during the
addition of new cuts; This has now been fixed.