Fix incorrect patch (35fe7dab8) for KaTeX overflow in quick search
The `position: relative` should be applied to each search result, not
on the searchResults container as a whole.

Otherwise, if there are 1-2 items (such that the total height of the
quick search box is less than the max height of searchResults), then
the overflow will still happen. To replicate this, you can search for
"Partialbruchzerlegung Berechnung" on production data with prior
commits.