Fix browser freezing when PostgreSQL decides not to highlight any matches
Frontend entered an infinite loop in MarkdownText.addMarks when PostgreSQL
decides to return search results from ts_query but doesn't highlight them
when requesting ts_headline with the same query. This seems to happen on
staging/prod databases and could not be replicated locally even when all
data was copied, which leads me to believe it's a locale issue (my laptop
is en_US for reference).

When answer/comment.highlighted_words from the backend is an empty array,
it gets passed as `addMarks(obj, new RegExp([].join("|")))`. The RegExp
instance matches all inputs if constructed with an empty string. This
results in the `while` loop within addMarks to never terminate.

This never happened on the old search page (and only happens on QuickSearch)
since the old search page uses a wrapper component `HighlightedMarkdown`
that checks for this exact case.

We fix the problem in two layers:
- When passing props to MarkdownText, QuickSearchResults checks if the
  array is empty and sets the regex prop to `undefined` instead
- Inside addMarks, if anyone accidentally does the same mistake in the
  future, we check if the RegExp is an empty match-all and return early.