Combine HighlightedMarkdown functionality into MarkdownText
This prevents any mis-use of MarkdownText directly in the future which
can lead to bugs like #386, caused by passing bad regex.

By limiting the scope of use to only highlighting an array of strings,
(AND checking whether the array is empty and acting accordingly), we
limit the blast radius for future developers.